move ProjectFilterContainer to new location, update code
