+++ common/components/componentsBySection/FindProjects/Filters/ProjectFilterContainer.jsx
@@ -4,13 +4,13 @@ import type {FluxReduceStore} from 'flux/utils';
 import {Container} from 'flux/utils';
 import React from 'react';
 import Select from 'react-select'
-import TagSelectorCollapsible from "../../common/tags/TagSelectorCollapsible.jsx";
-import TagCategory from "../../common/tags/TagCategory.jsx";
-import ProjectSearchDispatcher from '../../stores/ProjectSearchDispatcher.js';
-import ProjectSearchStore from '../../stores/ProjectSearchStore.js';
-import {Locations} from "../../constants/ProjectConstants";
-import {SelectOption} from "../../types/SelectOption.jsx";
-import metrics from "../../utils/metrics.js";
+import TagSelectorCollapsible from "../../../common/tags/TagSelectorCollapsible.jsx";
+import TagCategory from "../../../common/tags/TagCategory.jsx";
+import ProjectSearchDispatcher from '../../../stores/ProjectSearchDispatcher.js';
+import ProjectSearchStore from '../../../stores/ProjectSearchStore.js';
+import {Locations} from "../../../constants/ProjectConstants";
+import {SelectOption} from "../../../types/SelectOption.jsx";
+import metrics from "../../../utils/metrics.js";
 
 
 type State = {|
