+++ common/components/controllers/FindProjectsController.jsx
@@ -4,7 +4,7 @@ import ProjectSearchDispatcher from '../stores/ProjectSearchDispatcher.js';
 import TagDispatcher from '../stores/TagDispatcher.js';
 import ProjectCardsContainer from '../componentsBySection/FindProjects/ProjectCardsContainer.jsx';
 import ProjectSearchContainer from '../componentsBySection/FindProjects/ProjectSearchContainer.jsx';
-import ProjectFilterContainer from '../componentsBySection/FindProjects/ProjectFilterContainer.jsx';
+import ProjectFilterContainer from '../componentsBySection/FindProjects/Filters/ProjectFilterContainer.jsx';
 import {FindProjectsArgs} from "../stores/ProjectSearchStore.js";
 import SplashScreen from "../componentsBySection/FindProjects/SplashScreen.jsx";
 import urls from "../utils/url.js";
@@ -20,7 +20,7 @@ class FindProjectsController extends React.PureComponent<{||}, State> {
     super();
     this.state = {showSplash: true};
   }
-  
+
   componentWillMount(): void {
     let args: FindProjectsArgs = urls.arguments(document.location.search);
     args = _.pick(args, ['keyword','sortField','location','issues','tech', 'role', 'org', 'stage']);
@@ -28,7 +28,7 @@ class FindProjectsController extends React.PureComponent<{||}, State> {
     TagDispatcher.dispatch({type: 'INIT'});
     this.setState({showSplash: _.isEmpty(args)});
   }
-  
+
   _onClickFindProjects(): void {
     this.setState({showSplash: false});
   }
